requirements.txt
Flask>=2.3.0,<4.0.0
Pillow>=10.0.0,<11.0.0
pytesseract>=0.3.10,<0.4.0
openai>=1.0.0,<2.0.0
